big fixes & details screen ver 0.9.9.7

+++ app/src/main/java/com/example/niklas/efc_master/profiles/TempToPress.java
@@ -24,7 +24,6 @@ public final class TempToPress
 
 		int inX = 12;
 		for (; x < temp[inX]; inX--);
-			return (int)((press[inX]+((x-temp[inX])*(press[inX+1]-press[inX]) / (temp[inX+1] - temp[inX]))));
-			//return press[inX];
+			return (press[inX]+((x-temp[inX])*(press[inX+1]-press[inX]) / (temp[inX+1] - temp[inX])));
 	}
 }
